requirements.txt
python-dotenv
dbfread
psycopg2-binary
pandas